+++ test/setupTests.js
@@ -1,5 +1,6 @@
-const enzyme = require('enzyme');
-const Adapter = require('enzyme-adapter-react-16');
+import enzyme from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+
 import '@testing-library/jest-dom/extend-expect';
 import { configure } from '@testing-library/dom';