+++ Changelog.md
@@ -5,6 +5,7 @@
   - Limit array types to readonly where possible
   - Do not use void as a callback return value
 - Make linting fail if the type check fails
+- Remove conflicting ESLint rule 'quotes'
 
 ## 4.9.4
 ### Fixes